options button added as own component
