+++ src/components/post/PostHeader.js
@@ -1,12 +1,6 @@
 import React, { useState, useEffect } from "react";
 import styles from "../post/Post.module.css";
-import IconButton from "@mui/material/IconButton";
-import MoreHorizIcon from "@mui/icons-material/MoreHoriz";
-import Popover from "@mui/material/Popover";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
-import ListItemText from "@mui/material/ListItemText";
-import ButtonBase from "@mui/material/ButtonBase";
+import OptionsButton from "./OptionsButton";
 import ProfileImage from "./ProfileImage";
 import { useRouter } from "next/router";
 import { deletePost, followOrUnfollowUser } from "../../api";
@@ -25,13 +19,13 @@ export default function PostHeader({ post, currentUserId, updatePost }) {
     setIsFollowing(followerIds.includes(currentUserId));
   }, [post, currentUserId]);
 
-  const handleOptionsClick = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
+  // const handleOptionsClick = (event) => {
+  //   setAnchorEl(event.currentTarget);
+  // };
 
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
+  // const handleClose = () => {
+  //   setAnchorEl(null);
+  // };
 
   const handleDeletePost = async () => {
     const isDeleted = await deletePost(post.id);
@@ -44,10 +38,10 @@ export default function PostHeader({ post, currentUserId, updatePost }) {
   };
 
   const handleFollowOrUnfollowUser = async () => {
-    const authorId = post.author.id;
+    const postAuthorId = post.author.id;
 
     if (!isFollowing) {
-      const isFollowed = await followOrUnfollowUser(authorId, "follow");
+      const isFollowed = await followOrUnfollowUser(postAuthorId, "follow");
       if (isFollowed) {
         setIsFollowing(true);
         // Handle the success case, e.g., update the UI, show a notification, etc.
@@ -55,7 +49,7 @@ export default function PostHeader({ post, currentUserId, updatePost }) {
         // Handle the error case
       }
     } else {
-      const isUnfollowed = await followOrUnfollowUser(authorId, "unfollow");
+      const isUnfollowed = await followOrUnfollowUser(postAuthorId, "unfollow");
       if (isUnfollowed) {
         setIsFollowing(false);
         // Handle the success case, e.g., update the UI, show a notification, etc.
@@ -65,8 +59,8 @@ export default function PostHeader({ post, currentUserId, updatePost }) {
     }
   };
 
-  const open = Boolean(anchorEl);
-  const id = open ? "simple-popover" : undefined;
+  // const open = Boolean(anchorEl);
+  // const id = open ? "simple-popover" : undefined;
 
   return (
     <div className={styles.topPost}>
@@ -74,59 +68,14 @@ export default function PostHeader({ post, currentUserId, updatePost }) {
         <ProfileImage imageUrl={imageUrl} />
         <h3>{post.author.username}</h3>
       </div>
-      <div className={styles.optionsButton}>
-        <IconButton onClick={handleOptionsClick}>
-          <MoreHorizIcon />
-        </IconButton>
-        <Popover
-          id={id}
-          open={open}
-          anchorEl={anchorEl}
-          onClose={handleClose}
-          anchorOrigin={{
-            vertical: "bottom",
-            horizontal: "center",
-          }}
-          transformOrigin={{
-            vertical: "top",
-            horizontal: "center",
-          }}
-        >
-          <List component="nav">
-            <ListItem>
-              <ButtonBase
-                onClick={() => {
-                  router.push(`/users/${post.author.id}`);
-                }}
-              >
-                <ListItemText primary="Go to Profile" />
-              </ButtonBase>
-            </ListItem>
-
-            {post.author.id !== currentUserId && (
-              <ListItem>
-                <ButtonBase onClick={handleFollowOrUnfollowUser}>
-                  <ListItemText
-                    primary={
-                      isFollowing
-                        ? `Unfollow ${post.author.username}`
-                        : `Follow ${post.author.username}`
-                    }
-                  />
-                </ButtonBase>
-              </ListItem>
-            )}
-
-            {post.author.id === currentUserId && (
-              <ListItem>
-                <ButtonBase onClick={handleDeletePost}>
-                  <ListItemText primary="Delete post" />
-                </ButtonBase>
-              </ListItem>
-            )}
-          </List>
-        </Popover>
-      </div>
+      <OptionsButton
+        post={post}
+        currentUserId={currentUserId}
+        isFollowing={isFollowing}
+        handleFollowOrUnfollowUser={handleFollowOrUnfollowUser}
+        handleDeletePost={handleDeletePost}
+        router={router}
+      />
     </div>
   );
 }
